mono/pyproject.toml
[tool.poetry]
name = "python-experiment"
version = "0.1.0"
description = "A simple Python experiment with Poetry"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [
    {include = "src"},
    {include = "scripts"},
]

[tool.poetry.dependencies]
python = "^3.12"
click = "^8.1.3"  # Simple CLI library for command-line interfaces
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"  # Settings management with Pydantic
openai = "^1.4.0"
python-dotenv = "^1.0.0"
tenacity = "^8.2.2"  # Retry library for API calls
google-generativeai = "0.3.2"  # Google's Gemini API with fixed version
google-api-python-client = "^2.100.0"  # Google API client for Gmail
google-auth-httplib2 = "^0.2.0"  # HTTP transport for Google Auth
google-auth-oauthlib = "^1.1.0"  # OAuth2 flow for Google APIs
langchain = ">=0.0.267"
langchain-openai = ">=0.0.2"
langchain-community = ">=0.0.9"
langchain-core = ">=0.0.9"
sqlalchemy = "^2.0.23"
flask = "^2.3.3"  # Web framework
flask-restful = "^0.3.10"  # RESTful API extension for Flask
flask-cors = "^4.0.0"  # Cross-Origin Resource Sharing
gunicorn = "^21.2.0"  # WSGI HTTP Server for production deployment
flask-socketio = "^5.3.3"  # Flask-SocketIO for WebSocket support
mcp = "^1.0.0"  # Model Context Protocol SDK

[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
black = "^23.10.1"
isort = "^5.12.0"
mypy = "^1.6.1"
pytest = "^7.4.3"
types-flask-cors = "^5.0.0.20250413"

[tool.poetry.scripts]
hello = "scripts.hello:main"
gmail-crawler = "scripts.gmail_crawler:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py312"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false 